fix file names, rename modules
