+++ frontend/src/scenes/onboarding/home/sections/DiscoverInsight.tsx
