+++ frontend/src/scenes/onboarding/home/sections/TiledIconModule.tsx
