+++ frontend/src/scenes/onboarding/home/sections/static/analyze-conversions.png
