+++ frontend/src/scenes/onboarding/home/sections/static/analyzing-behavior.png
