+++ frontend/src/scenes/onboarding/home/sections/static/event-tracking-overview.png
