+++ frontend/src/scenes/onboarding/home/sections/static/measure-retention.png
