+++ frontend/src/scenes/onboarding/home/sections/static/roll-out-features.png
