+++ frontend/src/scenes/onboarding/home/sections/static/sales-revenue-tracking.png
