+++ frontend/src/scenes/onboarding/home/sections/static/tracking-b2b.png
