+++ frontend/src/scenes/onboarding/home/sections/static/tracking-spas.png
