+++ frontend/src/scenes/onboarding/home/sections/static/tracking-teams.png
